Update es-service-software.md
In my testing pressing the button to upgrade the software seems to get you in a queue but doesn't control *when* the install happens. So, just because I click the button to install after-hours doesn't mean the install will follow in that time frame. I.e., the install time cannot be controlled. So, removing this comment.